+++ src/prisma/prisma.module.ts
@@ -6,4 +6,5 @@ import { PrismaService } from './prisma.service.js';
   providers: [PrismaService],
   exports: [PrismaService],
 })
+// eslint-disable-next-line @typescript-eslint/no-extraneous-class -- NestJS @Module requires a class declaration
 export class PrismaModule {}